--- a/src/__tests__/basic-infrastructure.test.ts
+++ b/src/__tests__/basic-infrastructure.test.ts
@@ -1,0 +1,84 @@
+/**
+ * Basic Infrastructure Test
+ * Ensures the testing infrastructure is working correctly
+ */
+
+describe('Basic Infrastructure Tests', () => {
+  it('should run basic tests successfully', () => {
+    expect(true).toBe(true);
+  });
+
+  it('should have access to environment variables', () => {
+    expect(process.env.NODE_ENV).toBe('test');
+  });
+
+  it('should have Jest globals available', () => {
+    expect(typeof describe).toBe('function');
+    expect(typeof it).toBe('function');
+    expect(typeof expect).toBe('function');
+  });
+
+  it('should have mocked fetch available', () => {
+    expect(global.fetch).toBeDefined();
+    expect(typeof global.fetch).toBe('function');
+  });
+
+  it('should have crypto API available', () => {
+    expect(global.crypto).toBeDefined();
+    expect(typeof global.crypto.randomUUID).toBe('function');
+  });
+
+  it('should have TextEncoder/TextDecoder available', () => {
+    expect(global.TextEncoder).toBeDefined();
+    expect(global.TextDecoder).toBeDefined();
+  });
+
+  it('should handle async operations', async () => {
+    const result = await Promise.resolve('test');
+    expect(result).toBe('test');
+  });
+
+  it('should handle timeouts properly', (done) => {
+    setTimeout(() => {
+      expect(true).toBe(true);
+      done();
+    }, 10);
+  });
+});
+
+describe('Module Resolution Tests', () => {
+  it('should resolve @/ aliases', () => {
+    // This test ensures our module aliases are working
+    expect(() => {
+      // Just testing that the alias resolution doesn't throw
+      const path = '@/lib/config';
+      expect(path).toBeDefined();
+    }).not.toThrow();
+  });
+});
+
+describe('Mock Functionality Tests', () => {
+  it('should have localStorage mock', () => {
+    expect(window.localStorage).toBeDefined();
+    expect(typeof window.localStorage.getItem).toBe('function');
+    expect(typeof window.localStorage.setItem).toBe('function');
+  });
+
+  it('should have sessionStorage mock', () => {
+    expect(window.sessionStorage).toBeDefined();
+    expect(typeof window.sessionStorage.getItem).toBe('function');
+    expect(typeof window.sessionStorage.setItem).toBe('function');
+  });
+
+  it('should have ResizeObserver mock', () => {
+    expect(global.ResizeObserver).toBeDefined();
+    const observer = new global.ResizeObserver(() => {});
+    expect(typeof observer.observe).toBe('function');
+  });
+
+  it('should have IntersectionObserver mock', () => {
+    expect(global.IntersectionObserver).toBeDefined();
+    const observer = new global.IntersectionObserver(() => {});
+    expect(typeof observer.observe).toBe('function');
+  });
+});