--- a/src/__tests__/ci-cd-validation.test.ts
+++ b/src/__tests__/ci-cd-validation.test.ts
@@ -1,0 +1,210 @@
+/**
+ * CI/CD Pipeline Validation Tests
+ * Tests to ensure CI/CD pipeline components work correctly
+ */
+
+describe('CI/CD Pipeline Validation', () => {
+  describe('Test Infrastructure', () => {
+    it('should have Jest configured correctly', () => {
+      expect(typeof describe).toBe('function');
+      expect(typeof it).toBe('function');
+      expect(typeof expect).toBe('function');
+      expect(typeof beforeEach).toBe('function');
+      expect(typeof afterEach).toBe('function');
+    });
+
+    it('should have proper environment setup', () => {
+      expect(process.env.NODE_ENV).toBe('test');
+      expect(process.env.DATABASE_URL).toBeDefined();
+      expect(process.env.REDIS_URL).toBeDefined();
+      expect(process.env.JWT_SECRET).toBeDefined();
+      expect(process.env.ENCRYPTION_KEY).toBeDefined();
+    });
+
+    it('should have required globals available', () => {
+      expect(global.fetch).toBeDefined();
+      expect(global.TextEncoder).toBeDefined();
+      expect(global.TextDecoder).toBeDefined();
+      expect(global.crypto).toBeDefined();
+      expect(global.ResizeObserver).toBeDefined();
+      expect(global.IntersectionObserver).toBeDefined();
+    });
+  });
+
+  describe('Package Scripts Validation', () => {
+    it('should have all required test scripts defined', () => {
+      const packageJson = require('../../package.json');
+      const scripts = packageJson.scripts;
+
+      // Core test scripts
+      expect(scripts['test']).toBeDefined();
+      expect(scripts['test:unit']).toBeDefined();
+      expect(scripts['test:integration']).toBeDefined();
+      expect(scripts['test:e2e']).toBeDefined();
+      expect(scripts['test:security']).toBeDefined();
+      expect(scripts['test:performance']).toBeDefined();
+      expect(scripts['test:accessibility']).toBeDefined();
+
+      // Code quality scripts
+      expect(scripts['lint']).toBeDefined();
+      expect(scripts['format:check']).toBeDefined();
+      expect(scripts['type-check']).toBeDefined();
+
+      // Coverage scripts
+      expect(scripts['test:coverage']).toBeDefined();
+      expect(scripts['test:coverage:threshold']).toBeDefined();
+    });
+
+    it('should have build and development scripts', () => {
+      const packageJson = require('../../package.json');
+      const scripts = packageJson.scripts;
+
+      expect(scripts['dev']).toBeDefined();
+      expect(scripts['build']).toBeDefined();
+      expect(scripts['start']).toBeDefined();
+    });
+  });
+
+  describe('Dependencies Validation', () => {
+    it('should have core testing dependencies', () => {
+      const packageJson = require('../../package.json');
+      const devDeps = packageJson.devDependencies;
+
+      expect(devDeps['jest']).toBeDefined();
+      expect(devDeps['@testing-library/jest-dom']).toBeDefined();
+      expect(devDeps['@testing-library/react']).toBeDefined();
+      expect(devDeps['@playwright/test']).toBeDefined();
+      expect(devDeps['@axe-core/playwright']).toBeDefined();
+    });
+
+    it('should have code quality dependencies', () => {
+      const packageJson = require('../../package.json');
+      const devDeps = packageJson.devDependencies;
+
+      expect(devDeps['eslint']).toBeDefined();
+      expect(devDeps['prettier']).toBeDefined();
+      expect(devDeps['typescript']).toBeDefined();
+    });
+
+    it('should have security testing dependencies', () => {
+      const packageJson = require('../../package.json');
+      const devDeps = packageJson.devDependencies;
+
+      expect(devDeps['audit-ci']).toBeDefined();
+      expect(devDeps['snyk']).toBeDefined();
+    });
+  });
+
+  describe('Configuration Files Validation', () => {
+    it('should have Jest configuration', () => {
+      expect(() => require('../../jest.config.js')).not.toThrow();
+    });
+
+    it('should have Playwright configuration', () => {
+      expect(() => require('../../playwright.config.ts')).not.toThrow();
+    });
+
+    it('should have Next.js configuration', () => {
+      expect(() => require('../../next.config.js')).not.toThrow();
+    });
+
+    it('should have TypeScript configuration', () => {
+      const fs = require('fs');
+      const path = require('path');
+      const tsConfigPath = path.join(process.cwd(), 'tsconfig.json');
+      expect(fs.existsSync(tsConfigPath)).toBe(true);
+    });
+  });
+
+  describe('Test File Structure', () => {
+    it('should have basic test files', () => {
+      const fs = require('fs');
+      const path = require('path');
+
+      // Check for basic infrastructure test
+      const basicTestPath = path.join(__dirname, 'basic-infrastructure.test.ts');
+      expect(fs.existsSync(basicTestPath)).toBe(true);
+
+      // Check for E2E test directory
+      const e2eTestDir = path.join(process.cwd(), 'src/tests/e2e');
+      expect(fs.existsSync(e2eTestDir)).toBe(true);
+
+      // Check for accessibility test directory
+      const a11yTestDir = path.join(process.cwd(), 'src/tests/accessibility');
+      expect(fs.existsSync(a11yTestDir)).toBe(true);
+    });
+
+    it('should have test setup files', () => {
+      const fs = require('fs');
+      const path = require('path');
+
+      const jestSetupPath = path.join(process.cwd(), 'jest.setup.js');
+      expect(fs.existsSync(jestSetupPath)).toBe(true);
+
+      const globalSetupPath = path.join(process.cwd(), 'src/__tests__/setup/globalSetup.js');
+      expect(fs.existsSync(globalSetupPath)).toBe(true);
+
+      const globalTeardownPath = path.join(process.cwd(), 'src/__tests__/setup/globalTeardown.js');
+      expect(fs.existsSync(globalTeardownPath)).toBe(true);
+    });
+  });
+
+  describe('Script Files Validation', () => {
+    it('should have required script files', () => {
+      const fs = require('fs');
+      const path = require('path');
+
+      const scriptsDir = path.join(process.cwd(), 'scripts');
+      expect(fs.existsSync(scriptsDir)).toBe(true);
+
+      // Check for specific script files
+      const requiredScripts = [
+        'accessibility-test.js',
+        'memory-test.js',
+        'security-test.js',
+        'load-test.js'
+      ];
+
+      requiredScripts.forEach(script => {
+        const scriptPath = path.join(scriptsDir, script);
+        expect(fs.existsSync(scriptPath)).toBe(true);
+      });
+    });
+  });
+
+  describe('CI/CD Workflow Validation', () => {
+    it('should have GitHub Actions workflow', () => {
+      const fs = require('fs');
+      const path = require('path');
+
+      const workflowPath = path.join(process.cwd(), '.github/workflows/ci-cd.yml');
+      expect(fs.existsSync(workflowPath)).toBe(true);
+    });
+
+    it('should have proper workflow structure', () => {
+      const fs = require('fs');
+      const path = require('path');
+      const yaml = require('js-yaml');
+
+      const workflowPath = path.join(process.cwd(), '.github/workflows/ci-cd.yml');
+      const workflowContent = fs.readFileSync(workflowPath, 'utf8');
+      
+      expect(() => yaml.load(workflowContent)).not.toThrow();
+    });
+  });
+
+  describe('Error Handling', () => {
+    it('should handle test failures gracefully', () => {
+      // Test that our error handling works
+      expect(() => {
+        throw new Error('Test error');
+      }).toThrow('Test error');
+    });
+
+    it('should handle async errors', async () => {
+      await expect(async () => {
+        throw new Error('Async test error');
+      }).rejects.toThrow('Async test error');
+    });
+  });
+});