--- a/src/__tests__/performance/basic-performance.test.ts
+++ b/src/__tests__/performance/basic-performance.test.ts
@@ -1,0 +1,151 @@
+/**
+ * Basic Performance Tests
+ * Tests basic performance metrics and thresholds
+ */
+
+describe('Basic Performance Tests', () => {
+  beforeEach(() => {
+    // Reset performance marks
+    if (typeof performance !== 'undefined' && performance.clearMarks) {
+      performance.clearMarks();
+      performance.clearMeasures();
+    }
+  });
+
+  it('should complete basic operations within acceptable time', () => {
+    const start = Date.now();
+    
+    // Simulate some basic operations
+    const data = Array.from({ length: 1000 }, (_, i) => i);
+    const processed = data.map(x => x * 2).filter(x => x % 4 === 0);
+    
+    const end = Date.now();
+    const duration = end - start;
+    
+    expect(processed.length).toBeGreaterThan(0);
+    expect(duration).toBeLessThan(100); // Should complete in less than 100ms
+  });
+
+  it('should handle large data sets efficiently', () => {
+    const start = Date.now();
+    
+    // Create a larger dataset
+    const largeData = Array.from({ length: 10000 }, (_, i) => ({
+      id: i,
+      name: `Item ${i}`,
+      value: Math.random() * 100
+    }));
+    
+    // Perform operations
+    const filtered = largeData.filter(item => item.value > 50);
+    const sorted = filtered.sort((a, b) => b.value - a.value);
+    const top10 = sorted.slice(0, 10);
+    
+    const end = Date.now();
+    const duration = end - start;
+    
+    expect(top10.length).toBe(10);
+    expect(duration).toBeLessThan(500); // Should complete in less than 500ms
+  });
+
+  it('should handle async operations efficiently', async () => {
+    const start = Date.now();
+    
+    // Simulate async operations
+    const promises = Array.from({ length: 10 }, (_, i) => 
+      new Promise(resolve => setTimeout(() => resolve(i), 10))
+    );
+    
+    const results = await Promise.all(promises);
+    
+    const end = Date.now();
+    const duration = end - start;
+    
+    expect(results.length).toBe(10);
+    expect(duration).toBeLessThan(200); // Should complete in less than 200ms
+  });
+
+  it('should have efficient memory usage', () => {
+    // Test memory usage patterns
+    const initialMemory = process.memoryUsage();
+    
+    // Create and cleanup large objects
+    let largeArray = Array.from({ length: 100000 }, (_, i) => ({ id: i, data: `data-${i}` }));
+    
+    // Process the data
+    const processed = largeArray.map(item => item.id).filter(id => id % 2 === 0);
+    
+    // Cleanup
+    largeArray = null;
+    
+    if (global.gc) {
+      global.gc();
+    }
+    
+    const finalMemory = process.memoryUsage();
+    
+    expect(processed.length).toBeGreaterThan(0);
+    // Memory usage should not increase dramatically
+    const memoryIncrease = finalMemory.heapUsed - initialMemory.heapUsed;
+    expect(memoryIncrease).toBeLessThan(50 * 1024 * 1024); // Less than 50MB increase
+  });
+
+  it('should handle concurrent operations', async () => {
+    const start = Date.now();
+    
+    // Simulate concurrent operations
+    const concurrentTasks = Array.from({ length: 5 }, async (_, i) => {
+      const data = Array.from({ length: 1000 }, (_, j) => i * 1000 + j);
+      return data.reduce((sum, val) => sum + val, 0);
+    });
+    
+    const results = await Promise.all(concurrentTasks);
+    
+    const end = Date.now();
+    const duration = end - start;
+    
+    expect(results.length).toBe(5);
+    expect(results.every(result => typeof result === 'number')).toBe(true);
+    expect(duration).toBeLessThan(300); // Should complete in less than 300ms
+  });
+
+  it('should have acceptable function execution time', () => {
+    // Test function performance
+    const testFunction = (n: number) => {
+      let result = 0;
+      for (let i = 0; i < n; i++) {
+        result += Math.sqrt(i);
+      }
+      return result;
+    };
+
+    const start = performance.now();
+    const result = testFunction(10000);
+    const end = performance.now();
+    
+    const duration = end - start;
+    
+    expect(result).toBeGreaterThan(0);
+    expect(duration).toBeLessThan(50); // Should complete in less than 50ms
+  });
+
+  it('should handle string operations efficiently', () => {
+    const start = Date.now();
+    
+    // Test string operations
+    let text = '';
+    for (let i = 0; i < 1000; i++) {
+      text += `Item ${i} `;
+    }
+    
+    const words = text.split(' ').filter(word => word.length > 0);
+    const processed = words.map(word => word.toUpperCase()).join('|');
+    
+    const end = Date.now();
+    const duration = end - start;
+    
+    expect(words.length).toBe(1000);
+    expect(processed.length).toBeGreaterThan(0);
+    expect(duration).toBeLessThan(100); // Should complete in less than 100ms
+  });
+});