--- a/src/__tests__/security/basic-security.test.ts
+++ b/src/__tests__/security/basic-security.test.ts
@@ -1,0 +1,187 @@
+/**
+ * Basic Security Tests
+ * Tests basic security measures and configurations
+ */
+
+describe('Basic Security Tests', () => {
+  it('should have secure environment configuration', () => {
+    // Check that sensitive environment variables are not exposed
+    expect(process.env.NODE_ENV).toBe('test');
+    expect(process.env.JWT_SECRET).toBeDefined();
+    expect(process.env.ENCRYPTION_KEY).toBeDefined();
+  });
+
+  it('should handle input sanitization', () => {
+    // Test basic input sanitization
+    const maliciousInput = '<script>alert("xss")</script>';
+    const sanitized = maliciousInput.replace(/<script\b[^<]*(?:(?!<\/script>)<[^<]*)*<\/script>/gi, '');
+    
+    expect(sanitized).not.toContain('<script>');
+    expect(sanitized).not.toContain('alert');
+  });
+
+  it('should validate data types properly', () => {
+    // Test type validation
+    const validateEmail = (email: string): boolean => {
+      const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+      return emailRegex.test(email);
+    };
+
+    expect(validateEmail('test@example.com')).toBe(true);
+    expect(validateEmail('invalid-email')).toBe(false);
+    expect(validateEmail('')).toBe(false);
+    expect(validateEmail('test@')).toBe(false);
+  });
+
+  it('should handle SQL injection prevention patterns', () => {
+    // Test SQL injection prevention patterns
+    const sanitizeInput = (input: string): string => {
+      return input.replace(/['"\\;]/g, '');
+    };
+
+    const maliciousSQL = "'; DROP TABLE users; --";
+    const sanitized = sanitizeInput(maliciousSQL);
+    
+    expect(sanitized).not.toContain("'");
+    expect(sanitized).not.toContain('"');
+    expect(sanitized).not.toContain(';');
+    expect(sanitized).not.toContain('DROP');
+  });
+
+  it('should validate URL patterns', () => {
+    // Test URL validation
+    const validateURL = (url: string): boolean => {
+      try {
+        new URL(url);
+        return true;
+      } catch {
+        return false;
+      }
+    };
+
+    expect(validateURL('https://example.com')).toBe(true);
+    expect(validateURL('http://localhost:3000')).toBe(true);
+    expect(validateURL('invalid-url')).toBe(false);
+    expect(validateURL('javascript:alert(1)')).toBe(false);
+  });
+
+  it('should handle password security patterns', () => {
+    // Test password validation
+    const validatePassword = (password: string): boolean => {
+      const minLength = password.length >= 8;
+      const hasUpperCase = /[A-Z]/.test(password);
+      const hasLowerCase = /[a-z]/.test(password);
+      const hasNumbers = /\d/.test(password);
+      const hasSpecialChar = /[!@#$%^&*(),.?":{}|<>]/.test(password);
+      
+      return minLength && hasUpperCase && hasLowerCase && hasNumbers && hasSpecialChar;
+    };
+
+    expect(validatePassword('Password123!')).toBe(true);
+    expect(validatePassword('password')).toBe(false);
+    expect(validatePassword('PASSWORD')).toBe(false);
+    expect(validatePassword('12345678')).toBe(false);
+    expect(validatePassword('Pass1!')).toBe(false); // Too short
+  });
+
+  it('should handle CSRF token patterns', () => {
+    // Test CSRF token generation and validation
+    const generateCSRFToken = (): string => {
+      return Math.random().toString(36).substring(2, 15) + 
+             Math.random().toString(36).substring(2, 15);
+    };
+
+    const token1 = generateCSRFToken();
+    const token2 = generateCSRFToken();
+
+    expect(token1).toBeDefined();
+    expect(token2).toBeDefined();
+    expect(token1).not.toBe(token2); // Tokens should be unique
+    expect(token1.length).toBeGreaterThan(10);
+  });
+
+  it('should validate file upload security', () => {
+    // Test file upload validation
+    const validateFileType = (filename: string, allowedTypes: string[]): boolean => {
+      const extension = filename.split('.').pop()?.toLowerCase();
+      return extension ? allowedTypes.includes(extension) : false;
+    };
+
+    const allowedTypes = ['jpg', 'jpeg', 'png', 'gif', 'pdf'];
+
+    expect(validateFileType('image.jpg', allowedTypes)).toBe(true);
+    expect(validateFileType('document.pdf', allowedTypes)).toBe(true);
+    expect(validateFileType('script.js', allowedTypes)).toBe(false);
+    expect(validateFileType('malware.exe', allowedTypes)).toBe(false);
+  });
+
+  it('should handle rate limiting patterns', () => {
+    // Test rate limiting logic
+    const rateLimiter = {
+      requests: new Map<string, number[]>(),
+      isAllowed(ip: string, maxRequests: number = 10, windowMs: number = 60000): boolean {
+        const now = Date.now();
+        const requests = this.requests.get(ip) || [];
+        
+        // Remove old requests outside the window
+        const validRequests = requests.filter(time => now - time < windowMs);
+        
+        if (validRequests.length >= maxRequests) {
+          return false;
+        }
+        
+        validRequests.push(now);
+        this.requests.set(ip, validRequests);
+        return true;
+      }
+    };
+
+    const testIP = '192.168.1.1';
+    
+    // Should allow first few requests
+    expect(rateLimiter.isAllowed(testIP, 3)).toBe(true);
+    expect(rateLimiter.isAllowed(testIP, 3)).toBe(true);
+    expect(rateLimiter.isAllowed(testIP, 3)).toBe(true);
+    
+    // Should block after limit
+    expect(rateLimiter.isAllowed(testIP, 3)).toBe(false);
+  });
+
+  it('should validate session security', () => {
+    // Test session validation patterns
+    const validateSession = (sessionData: any): boolean => {
+      if (!sessionData || typeof sessionData !== 'object') {
+        return false;
+      }
+      
+      const requiredFields = ['userId', 'createdAt', 'expiresAt'];
+      const hasRequiredFields = requiredFields.every(field => sessionData[field]);
+      
+      if (!hasRequiredFields) {
+        return false;
+      }
+      
+      const now = Date.now();
+      const isNotExpired = sessionData.expiresAt > now;
+      
+      return isNotExpired;
+    };
+
+    const validSession = {
+      userId: '123',
+      createdAt: Date.now() - 1000,
+      expiresAt: Date.now() + 3600000 // 1 hour from now
+    };
+
+    const expiredSession = {
+      userId: '123',
+      createdAt: Date.now() - 3600000,
+      expiresAt: Date.now() - 1000 // Expired
+    };
+
+    expect(validateSession(validSession)).toBe(true);
+    expect(validateSession(expiredSession)).toBe(false);
+    expect(validateSession(null)).toBe(false);
+    expect(validateSession({})).toBe(false);
+  });
+});